refactor(apps-api-graph): revert to string args for typeorm
