+++ apps/api/graph/src/app/app.module.ts
@@ -2,9 +2,10 @@ import { Module, OnModuleInit } from '@nestjs/common'
 import { ConfigService } from '@nestjs/config'
 import { GraphQLModule } from '@nestjs/graphql'
 import { TypeOrmModule } from '@nestjs/typeorm'
-import * as shell from 'shelljs'
-import { EdgeModule, GraphModule, UserModule, VertexModule } from '../models'
-import { SeedDbService } from '../seed-db'
+import { EdgeModule } from '../models/edge/edge.module'
+import { GraphModule } from '../models/graph/graph.module'
+import { UserModule } from '../models/user/user.module'
+import { VertexModule } from '../models/vertex/vertex.module'
 import { AppController } from './app.controller'
 import { AppService } from './app.service'
 import {
@@ -39,22 +40,21 @@ import {
   providers: [AppService],
 })
 export class AppModule implements OnModuleInit {
-  constructor(public seedDbService: SeedDbService) {}
+  // constructor(public seedDbService: SeedDbService) {}
 
   async onModuleInit() {
-    if (process.argv.includes('--reset')) {
-      if (
-        shell.exec(
-          `npx hasura metadata apply \
-          --project apps/api/graph/.hasura \
-          --envfile .env`,
-        ).code !== 0
-      ) {
-        shell.echo('"hasura metadata apply" failed')
-        shell.exit(1)
-      }
-
-      await this.seedDbService.seedDB()
-    }
+    // if (process.argv.includes('--reset')) {
+    //   if (
+    //     shell.exec(
+    //       `npx hasura metadata apply \
+    //       --project apps/api/graph/.hasura \
+    //       --envfile .env`,
+    //     ).code !== 0
+    //   ) {
+    //     shell.echo('"hasura metadata apply" failed')
+    //     shell.exit(1)
+    //   }
+    //   await this.seedDbService.seedDB()
+    // }
   }
 }
