--- apps/api/graph/src/models/edge/index.ts
@@ -1,2 +1,0 @@
-export * from './edge.module'
-export * from './edge.entity'
