+++ apps/api/graph/src/models/graph/graph.entity.ts
@@ -1,8 +1,8 @@
 import { ObjectType } from '@nestjs/graphql'
 import { Entity, ManyToOne, OneToMany, PrimaryGeneratedColumn } from 'typeorm'
-import { EdgeEntity } from '../edge'
-import { UserEntity } from '../user'
-import { VertexEntity } from '../vertex'
+import { EdgeEntity } from '../edge/edge.entity'
+import { UserEntity } from '../user/user.entity'
+import { VertexEntity } from '../vertex/vertex.entity'
 import { IGraph } from './IGraph'
 
 @Entity('graph')
