--- apps/api/graph/src/models/graph/index.ts
@@ -1,2 +1,0 @@
-export * from './graph.entity'
-export * from './graph.module'
