--- apps/api/graph/src/models/index.ts
@@ -1,4 +1,0 @@
-export * from './edge'
-export * from './graph'
-export * from './user'
-export * from './vertex'
