--- apps/api/graph/src/models/user/index.ts
@@ -1,5 +1,0 @@
-export * from './IUser'
-export * from './user.entity'
-export * from './user.module'
-export * from './user.resolver'
-export * from './user.service'
