--- apps/api/graph/src/models/vertex/index.ts
@@ -1,2 +1,0 @@
-export * from './vertex.module'
-export * from './vertex.entity'
