--- apps/api/graph/src/seed-db/index.ts
@@ -1,2 +1,0 @@
-export * from './seed-db.module'
-export * from './seed-db.service'
