+++ apps/api/graph/src/seed-db/seed-db.module.ts
@@ -1,5 +1,8 @@
 import { Module } from '@nestjs/common'
-import { EdgeModule, GraphModule, UserModule, VertexModule } from '../models'
+import { EdgeModule } from '../models/edge/edge.module'
+import { GraphModule } from '../models/graph/graph.module'
+import { UserModule } from '../models/user/user.module'
+import { VertexModule } from '../models/vertex/vertex.module'
 import { SeedDbService } from './seed-db.service'
 
 @Module({
